updated README and INSTALL instructions
README.md now includes a section for frontend tests, and notes that local install is needed for tests.
INSTALL.md has been updated to include correct instructions for locally installing/running the backend
and frontend services.
updated the recommendation for postgresql to run on a docker container instead of locally.